cloud_functions/extract_images/requirements.txt
functions-framework==3.*
requests==2.28.2
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0